+++ spec/active_interaction/filters/boolean_filter_spec.rb
@@ -7,35 +7,19 @@ describe ActiveInteraction::BooleanFilter, :filter do
   it_behaves_like 'a filter'
 
   describe '#cast' do
-    context 'with false' do
-      let(:value) { false }
-
-      it 'returns false' do
-        expect(filter.cast(value)).to be_false
+    context 'falsey' do
+      [false, '0', 'false', 'FALSE'].each do |value|
+        it "returns false for #{value.inspect}" do
+          expect(filter.cast(value)).to be_false
+        end
       end
     end
 
-    context 'with true' do
-      let(:value) { true }
-
-      it 'returns true' do
-        expect(filter.cast(value)).to be_true
-      end
-    end
-
-    context 'with "0"' do
-      let(:value) { '0' }
-
-      it 'returns false' do
-        expect(filter.cast(value)).to be_false
-      end
-    end
-
-    context 'with "1"' do
-      let(:value) { '1' }
-
-      it 'returns true' do
-        expect(filter.cast(value)).to be_true
+    context 'truthy' do
+      [true, '1', 'true', 'TRUE'].each do |value|
+        it "returns true for #{value.inspect}" do
+          expect(filter.cast(value)).to be_true
+        end
       end
     end
   end